+++ groovy-android-gradle-plugin/src/main/groovy/groovyx/GroovyAndroidPlugin.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 the original author or authors.
+ * Copyright 2017 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -24,6 +24,7 @@ import com.android.build.gradle.internal.variant.BaseVariantData
 import com.android.build.gradle.internal.variant.BaseVariantOutputData
 import com.android.builder.model.SourceProvider
 import groovy.transform.CompileStatic
+import groovy.util.logging.Slf4j
 import org.gradle.api.GradleException
 import org.gradle.api.Plugin
 import org.gradle.api.Project
@@ -40,6 +41,7 @@ import javax.inject.Inject
 /**
  * Adds support for building Android applications using the Groovy language.
  */
+@Slf4j
 class GroovyAndroidPlugin implements Plugin<Project> {
 
   public static final String ANDROID_GROOVY_EXTENSION_NAME = 'androidGroovy'
@@ -72,6 +74,7 @@ class GroovyAndroidPlugin implements Plugin<Project> {
       def sourceSetPath = project.file("src/$sourceSetName/groovy")
 
       if (!sourceSetPath.exists()) {
+        log.debug('SourceSet path does not exists for {} {}', sourceSetName, sourceSetPath)
         return
       }
 
@@ -84,7 +87,7 @@ class GroovyAndroidPlugin implements Plugin<Project> {
       def groovyDirSet = groovySourceSet.groovy
       groovyDirSet.srcDir(sourceSetPath)
 
-      project.logger.debug("Created groovy sourceDirectorySet at $groovyDirSet.srcDirs")
+      log.debug('Created groovy sourceDirectorySet at {}', groovyDirSet.srcDirs)
     }
 
     project.afterEvaluate { Project afterProject ->
@@ -100,11 +103,11 @@ class GroovyAndroidPlugin implements Plugin<Project> {
 
     variantDataList.each { variantData ->
       def variantDataName = variantData.name
-      project.logger.debug("Process variant [$variantDataName]")
+      log.debug('Process variant {}', variantDataName)
 
       def javaTask = getJavaTask(variantData)
       if (javaTask == null) {
-        project.logger.info("javaTask is missing for $variantDataName, so Groovy files won't be compiled for it")
+        log.info('javaTask is missing for {}, so Groovy files won\'t be compiled for it', variantDataName)
         return
       }
 
@@ -126,8 +129,6 @@ class GroovyAndroidPlugin implements Plugin<Project> {
       providers.each { SourceProvider provider ->
         def groovySourceSet = provider.convention.plugins['groovy'] as GroovySourceSet
         if (groovySourceSet == null) {
-          // no source set skip task for this set
-          project.tasks.remove(groovyTask)
           return
         }
 
@@ -136,8 +137,6 @@ class GroovyAndroidPlugin implements Plugin<Project> {
 
         // Exclude any java files that may be included in both java and groovy source sets
         javaTask.exclude { file ->
-          project.logger.debug("Exclude java file $file.file")
-          project.logger.debug("Eexlude against groovy files $groovySourceSet.groovy.files")
           file.file in groovySourceSet.groovy.files
         }
 
@@ -149,14 +148,28 @@ class GroovyAndroidPlugin implements Plugin<Project> {
         }
       }
 
+      // no sources for groovy to compile skip the groovy task
+      if (groovyTask.source.empty) {
+        log.debug('no groovy sources found for {} removing groovy task', variantDataName)
+        project.tasks.remove(groovyTask)
+        return
+      }
+      log.debug('groovy sources for {}: {}', variantDataName, groovyTask.source.files)
+
       def additionalSourceFiles = getGeneratedSourceDirs(variantData)
       groovyTask.source(*additionalSourceFiles)
 
       groovyTask.doFirst { GroovyCompile task ->
         def androidRunTime = project.files(getRuntimeJars(androidPlugin, androidExtension))
-        task.classpath = androidRunTime + javaTask.classpath
+        def kotlinFiles = project.files(getKotlinClassFiles(variantDataName))
+        task.classpath = androidRunTime + javaTask.classpath + kotlinFiles
         task.groovyClasspath = task.classpath
+        task.options.compilerArgs += javaTask.options.compilerArgs
+        task.groovyOptions.javaAnnotationProcessing = true
       }
+
+      log.debug('Groovy classpath: {}', groovyTask.classpath.files)
+      log.debug('Groovy compiler args {}', groovyTask.options.compilerArgs)
 
       javaTask.finalizedBy(groovyTask)
     }
@@ -178,7 +191,9 @@ class GroovyAndroidPlugin implements Plugin<Project> {
   private static getRuntimeJars(BasePlugin plugin, BaseExtension extension) {
     if (plugin.metaClass.getMetaMethod('getRuntimeJarList')) {
       return plugin.runtimeJarList
-    } else if (extension.metaClass.getMetaMethod('getBootClasspath')) {
+    }
+
+    if (extension.metaClass.getMetaMethod('getBootClasspath')) {
       return extension.bootClasspath
     }
 
@@ -188,9 +203,12 @@ class GroovyAndroidPlugin implements Plugin<Project> {
   private static SourceTask getJavaTask(BaseVariantData baseVariantData) {
     if (baseVariantData.metaClass.getMetaProperty('javaCompileTask')) {
       return baseVariantData.javaCompileTask
-    } else if (baseVariantData.metaClass.getMetaProperty('javaCompilerTask')) {
+    }
+
+    if (baseVariantData.metaClass.getMetaProperty('javaCompilerTask')) {
       return baseVariantData.javaCompilerTask
     }
+
     return null
   }
 
@@ -230,6 +248,10 @@ class GroovyAndroidPlugin implements Plugin<Project> {
       result << variantData.scope.aidlSourceOutputDir
     }
 
+    if (variantData.scope.annotationProcessorOutputDir != null) {
+      result << variantData.scope.annotationProcessorOutputDir
+    }
+
     // We use getter instead of property for globalScope since property returns
     // TransformGlobalScope instead of GlobalScope (Static type checker failing).
     if (variantData.scope.getGlobalScope().extension.dataBinding.enabled) {
@@ -250,5 +272,17 @@ class GroovyAndroidPlugin implements Plugin<Project> {
 
   private static VariantManager getVariantManager(BasePlugin plugin) {
     return plugin.variantManager
+  }
+
+  private getKotlinClassFiles(String variantName) {
+    // kotlin plugin exists
+    def kotlin = project.plugins.findPlugin('kotlin-android')
+    if (kotlin != null) {
+      def kotlinCopyTask = project.tasks.findByName("copy${variantName.capitalize()}KotlinClasses")
+      return kotlinCopyTask.kotlinOutputDir
+    }
+
+    // empty list to avoid null issues when adding to FileCollection
+    return Collections.emptyList()
   }
 }